Add arg asserts to copyForwards and copyBackwards
Helps document in code what the comments already say. I had to fixup
a comptime use case of these functions in Encoding.zig (as you can't
retrieve the pointer value at comptime...). However, @memcpy seems to
work at comptime now so it resolves a pre-existing TODO there anyway!

It does raise the question of how you should handle comptime memcpys
with overlapping arrays though - as ideally you would be able to still
validate the ptr ordering in that case as well? (currently that is not
possible within the zig language).

This commit also fixes a bug where `@setRuntimeSafety` was being
called in the same scope as the asserts in `copyBackwards`, rendering
them useless!

Co-authored-by: rohlem <rohlemF@gmail.com>
Co-authored-by: Andrew Kelley <andrew@ziglang.org>