std.io: add AnyWriter
This is a non-generic writer that holds a type-erased pointer to a
generic writer and is suitable for use in scenarios, such as runtime
callbacks, where generic functions cannot be used.

The approach taken here is a bit different from AnyReader in that the
type-erased version has not replaced the generic version. Instead, an
interface mixin is used to share the same interface between the two
implementations.