Curve25519.fromEdwards25519(): don't assume normalized coordinates
The low-level `Curve25519.fromEdwards25519()` function assumed
that the X/Y coordinates were not scaled (t=1).

But this is not guaranteed to be the case.

In most real-world applications, the coordinates are freshly decoded,
either directly or via the X25519.fromEd25519() function, so this
is not an issue.

However, since we offer the ability to do that conversion after
arbitrary computations, the assertion was not correct.