BoundedArray: add a len() function to get the length as a usize
The type of the active length was changed to the smallest possible
type, which is nice to get a more compact representation.

However, that change introduced unexpected side effects.
For example, `a.len + b.len` can now trigger an integer overflow.

There's also an inconsistency between functions that set the size,
which all use a `usize`, and the way to read the size, that uses a
different type. This is also inconsistent with pretty much anything
else that represents a slice length.

Replace `.len` with `len()`, a function that returns the length as
a `usize` to minimize surprise and simplify application code.